Installed reselect and decomposed the reduce
