+++ src/index.js
@@ -3,7 +3,8 @@ import ReactDom from "react-dom";
 import {applyMiddleware, compose, createStore} from "redux";
 import thunk from "redux-thunk";
 import {Provider} from "react-redux";
-import {reducer, loadOffers} from "./reducer";
+import reducer from "./reducer/reducer.js";
+import {Operation as DataOperation} from "./reducer/data/data.js";
 import App from "./components/app/app.jsx";
 import {createApi} from "./api.js";
 
@@ -17,7 +18,7 @@ const store = createStore(
     )
 );
 
-store.dispatch(loadOffers());
+store.dispatch(DataOperation.loadData());
 
 ReactDom.render(
     <Provider store={store}>
